# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chinook API",
        "contact": {
            "name": "Testsmith",
            "email": "info@testsmith.io"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local environment"
        }
    ],
    "paths": {
...
        "/invoicelines": {
            "get": {
                "tags": [
                    "Invoiceline"
                ],
                "summary": "Retrieve all invoicelines",
                "operationId": "get-all-invoicelines",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page, default=1",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All invoiceline",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InvoicelineResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Invoiceline"
                ],
                "summary": "New invoicelines",
                "operationId": "store-invoiceline",
                "requestBody": {
                    "description": "Invoiceline object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvoicelineRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An Album",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoicelineResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/invoicelines/{invoicelineId}": {
            "get": {
                "tags": [
                    "Invoiceline"
                ],
                "summary": "Retrieve a specific invoiceline",
                "operationId": "get-invoiceline",
                "parameters": [
                    {
                        "name": "invoicelineId",
                        "in": "path",
                        "description": "The invoicelineId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An invoiceline",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoicelineResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            },
            "put": {
                "tags": [
                    "Invoiceline"
                ],
                "summary": "Update invoiceline",
                "operationId": "update-invoiceline",
                "parameters": [
                    {
                        "name": "invoicelineId",
                        "in": "path",
                        "description": "The invoicelineId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Invoiceline object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvoicelineRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An invoiceline",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InvoicelineResponse"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Invoiceline"
                ],
                "summary": "Remove invoiceline",
                "operationId": "destroy-invoiceline",
                "parameters": [
                    {
                        "name": "invoicelineId",
                        "in": "path",
                        "description": "The invoicelineId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InvoicelineRequest": {
                "title": "InvoicelineRequest",
                "required": [
                    "title",
                    "body"
                ],
                "properties": {
                    "invoice_id": {
                        "description": "Class Invoiceline",
                        "type": "integer"
                    },
                    "track_id": {
                        "description": "Class Invoiceline",
                        "type": "integer"
                    },
                    "unit_price": {
                        "description": "Class Invoiceline",
                        "type": "string"
                    },
                    "quantity": {
                        "description": "Class Invoiceline",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "InvoicelineResponse": {
                "title": "InvoicelineResponse",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "invoice_id": {
                        "type": "integer"
                    },
                    "track_id": {
                        "type": "integer"
                    },
                    "unit_price": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}